Melhoria no worker, agora ele possui a capacidade de processar varias residencias, porem o tempo de cada ciclo precisou ser maior, caso tenha um volume muito grande de residencias ele tera tempo suficiente para processar todas, adicionado algumas validações nas structs, e adicionado o endpoint de update, e agora o json de resposta tras o id do proprietario.

+++ internal/usecase/dto/dto.go
@@ -9,7 +9,7 @@ type WaterMonitoringInput struct {
 }
 
 type WaterMonitoringOutput struct {
-	ID           int    `json:"id"`
+	ID           int64  `json:"id"`
 	OwnerName    string `json:"owner"`
 	Email        string `json:"email"`
 	Phone        string `json:"phone"`
@@ -19,4 +19,5 @@ type WaterMonitoringOutput struct {
 	State        string `json:"state"`
 	City         string `json:"city"`
 	Region       string `json:"region"`
+	
 }
